backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
motor>=3.3.0
python-dotenv>=1.0.0
httpx>=0.24.0
pydantic>=2.0.0
email-validator>=2.0.0
pydantic-settings>=2.0.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
